+++ src/extension.ts
@@ -1,4 +1,4 @@
-import { LanguageClient, ServerOptions, LanguageClientOptions, StreamInfo } from 'vscode-languageclient'
+import { LanguageClient, ServerOptions, LanguageClientOptions, StreamInfo } from 'vscode-languageclient/node'
 import { EvaluatableExpressionRequest } from './protocol'
 
 import * as vscode from 'vscode'
@@ -23,7 +23,7 @@ interface PhpactorConfig {
     launchServerArgs: string[]
 }
 
-export function activate(context: vscode.ExtensionContext): void {
+export async function activate(context: vscode.ExtensionContext): Promise<void> {
     if (!checkPlatform()) {
         return
     }
@@ -50,7 +50,7 @@ export function activate(context: vscode.ExtensionContext): void {
     if (enable === false) return
 
     languageClient = createClient(config)
-    languageClient.start()
+    await languageClient.start()
 
     context.subscriptions.push(
         vscode.languages.registerEvaluatableExpressionProvider('php', {